# docs/api/OPENAPI_SPECIFICATION.yaml
openapi: 3.0.3
info:
  title: PsychSync API
  description: |
    PsychSync is a psychological assessment SaaS platform that provides personality assessments,
    team analytics, and AI-powered insights for organizations.

    **Authentication:**
    - All endpoints require JWT authentication unless marked as `public`
    - Include token in Authorization header: `Bearer <token>`
    - Token refresh endpoint available for obtaining new access tokens

    **Rate Limiting:**
    - Rate limits vary by endpoint (see specific endpoint documentation)
    - Rate limit headers included in all responses:
      - `X-RateLimit-Limit`: Max requests per window
      - `X-RateLimit-Remaining`: Remaining requests
      - `X-RateLimit-Reset`: Reset time (Unix timestamp)

    **Response Format:**
    - Success responses follow the schema defined in each endpoint
    - Error responses follow standard error format

    **Support:**
    - API Documentation: https://docs.psychsync.com/api
    - Support Email: api-support@psychsync.com
    - Status Page: https://status.psychsync.com
  version: 1.0.0
  contact:
    name: PsychSync API Team
    email: api-support@psychsync.com
    url: https://www.psychsync.com
  license:
    name: Proprietary
    url: https://www.psychsync.com/terms
  termsOfService: https://www.psychsync.com/terms
servers:
  - url: https://api.psychsync.com/api/v1
    description: Production server
  - url: https://api-staging.psychsync.com/api/v1
    description: Staging server
...
paths:
...
  /assessments/:
    get:
      tags:
        - Assessments
      summary: List assessments
      description: Retrieve a paginated list of assessments with optional filtering. Supports filtering by category, status, and search.
      operationId: listAssessments
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: limit
          in: query
          description: Number of results per page
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 1000
        - name: category
          in: query
          description: Filter by assessment category
          schema:
            type: string
            enum: [PERSONALITY, SKILLS, BEHAVIORAL, COGNITIVE]
        - name: status
          in: query
          description: Filter by assessment status
          schema:
            type: string
            enum: [DRAFT, PUBLISHED, ARCHIVED]
        - name: search
          in: query
          description: Search in title and description
          schema:
            type: string
        - name: sort
          in: query
          description: Field to sort by
          schema:
            type: string
            enum: [created_at, updated_at, title]
            default: created_at
        - name: order
          in: query
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: List of assessments
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Assessment'
                  total:
                    type: integer
                    description: Total number of assessments
                  page:
                    type: integer
                  limit:
                    type: integer
                  pages:
                    type: integer
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - Assessments
      summary: Create assessment
      description: Create a new assessment. User must be active and authenticated.
      operationId: createAssessment
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssessmentCreate'
      responses:
        '201':
          description: Assessment created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assessment'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from /auth/token-fixed endpoint
  schemas:
    Error:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Human-readable error message
        error_code:
          type: string
          description: Machine-readable error code
          example: VALIDATION_ERROR
        status_code:
          type: integer
          description: HTTP status code
          example: 400
        request_id:
          type: string
          format: uuid
          description: Unique request ID for troubleshooting
...
    Assessment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [PERSONALITY, SKILLS, BEHAVIORAL, COGNITIVE]
        status:
          type: string
          enum: [DRAFT, PUBLISHED, ARCHIVED]
        framework_code:
          type: string
          example: BIG_FIVE
        created_by_id:
          type: string
          format: uuid
        team_id:
          type: string
          format: uuid
          nullable: true
        organization_id:
          type: string
          format: uuid
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AssessmentCreate:
      type: object
      required:
        - title
        - category
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [PERSONALITY, SKILLS, BEHAVIORAL, COGNITIVE]
        framework_code:
          type: string
          description: Psychological framework to use
          example: BIG_FIVE
        team_id:
          type: string
          format: uuid
          description: Optional team association
...